requirements.txt
flask
openai
pymongo
requests
setuptools
dnspython
python-Levenshtein
unidecode